requirements.txt
flet
langstream
llama_cpp_python